requirements.txt
attrs==22.2.0
beautifulsoup4==4.11.1
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pycryptodome==3.21.0
pytest==7.2.1
python-dotenv==0.21.0
soupsieve==2.3.2.post1
tomli==2.0.1
pytest-check==2.1.2
PyYAML~=6.0.1
jsonschema~=4.19.1
lxml~=5.3.0
deepdiff~=8.5.0